Fix "naoth restart" to actually work.
The old code started itself again with start/stop parameters using the

$0 stop

syntax. This starts a child process and there are now two processes with the name "naoth", the parent and the child. When the stalled naoth processes are removed the "pidof" command is called with the parameter to ignore the one naoth process itself. It will still return the PID of the parent process (which also has the name "naoth" and thus the following "Kill" will kill the parent process and the whole script itself.

The fix is to use sub-functions in the script for start/stop instead of calling the script itself.