Don't use actual blocking device for camera synchronization but polling.
If we use blocking, we cannot clear the queue of images in case of frame lost. Thus we will work on older images (at most the number of buffers 5). By polling, we can still synchronize to the camera, but clear the queue of images, thus always getting the last image.