Adding minimal clang format configuration.
This sets the things we explicitly defined (intendation, basic brace placement and pointer alignment) but derives from Google Style for everything else as fallback.
It is not a decision to use Google Style!!! Whenever we decide on something regarding coding style that is not handled the same way as Google style, just override the configuration in this file (e.g. we are using 4 spaces instead of 2).

For a list of all configuration variables see http://clang.llvm.org/docs/ClangFormatStyleOptions.html#configurable-format-style-options